# docs/swagger.yaml
definitions:
  contact.ContactSubmissionInput:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
      subject:
        type: string
    required:
    - email
    - message
    - subject
    type: object
...
info:
  contact: {}
paths:
  /artikel:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua artikel
      parameters:
      - description: ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan semua artikel
      tags:
      - Artikel
    post:
      consumes:
      - multipart/form-data
      description: Menambahkan entri artikel baru
      parameters:
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Nama lengkap
        in: formData
        name: full_name
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Topik
        in: formData
        name: topic
        required: true
        type: string
      - description: Pesan artikel
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Menambahkan artikel baru
      tags:
      - Artikel
  /artikel/{slug}:
    delete:
      consumes:
      - application/json
      description: Menghapus artikel by id
      parameters:
      - description: Artikel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Menghapus artikel by id
      tags:
      - Artikel
    get:
      consumes:
      - application/json
      description: Mendapatkan satu artikel by id
      parameters:
      - description: Artikel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Mendapatkan satu artikel by id
      tags:
      - Artikel
    put:
      consumes:
      - multipart/form-data
      description: Memperbarui artikel dengan informasi yang diberikan
      parameters:
      - description: Artikel ID
        in: path
        name: id
        required: true
        type: integer
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Nama lengkap
        in: formData
        name: full_name
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Topik
        in: formData
        name: topic
        required: true
        type: string
      - description: Pesan artikel
        in: formData
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui artikel
      tags:
      - Artikel
  /contact:
    get:
      consumes:
      - application/json
      description: Get all contact form submissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get All Contact Submissions
      tags:
      - Contact
    post:
      consumes:
      - application/json
      description: Submit a contact form
      parameters:
      - description: Contact form input
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/contact.ContactSubmissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Submit Contact Form
      tags:
      - Contact
  /contact/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a contact form submission by ID
      parameters:
      - description: Contact Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete Contact Submission by ID
      tags:
      - Contact
    get:
      consumes:
      - application/json
      description: Get a contact form submission by ID
      parameters:
      - description: Contact Submission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get Contact Submission by ID
      tags:
      - Contact
...
swagger: "2.0"